Move hex utils to new fizzy::test-utils library
