+++ test/CMakeLists.txt
@@ -1,1 +1,2 @@
+add_subdirectory(utils)
 add_subdirectory(unittests)
