+++ test/unittests/end_to_end_test.cpp
@@ -1,7 +1,7 @@
 #include "execute.hpp"
-#include "hex.hpp"
 #include "parser.hpp"
 #include <gtest/gtest.h>
+#include <test/utils/hex.hpp>
 
 using namespace fizzy;
