+++ test/unittests/leb128_test.cpp
@@ -1,6 +1,6 @@
-#include "hex.hpp"
 #include "leb128.hpp"
 #include <gtest/gtest.h>
+#include <test/utils/hex.hpp>
 
 using namespace fizzy;
